Tests/LibWeb/Text/expected/css/font-face-set-load-filters-by-text.txt

MultiFont face count: 3
range=U+41 status=loaded
range=U+42 status=error
range=U+43 status=error
